tests: use library-provided AES key and block size macros
The file test_aes.c uses NUM_OF_NIST_KEYS for both the size of an AES
key and the size of an AES block. The two happen to be the same since
tinycrypt hard-codes AES-128, but that should not be assumed.

To fix that, replace NUM_OF_NIST_KEYS with either TC_AES_BLOCK_SIZE or
TC_AES_KEY_SIZE, depending on the case.

Signed-off-by: Marti Bolivar <marti.bolivar@linaro.org>