Fix innocuous memset size bug in ECC test method
In one of the ECC auxiliary test methods, a buffer 'd' was being
erased with memset using a wrong buffer size (NUM_ECC_WORDS instead
of 4*NUM_ECC_WORDS). This bug was not affecting the correctness of
ECC nor any other crypto primtive. In fact, it was not affecting even
the ECC test itself since the content of such buffer was being
overwritten rigth after the erasure procedure.

This fix removes a warning triggered when compiling the code.